moved testing functions out of mp_timeout_decorator
