+++ tests/unit/file_cache/mp_timeout_decorator_test.py
@@ -8,8 +8,10 @@ import polars as pl  # type: ignore
 from pl_series_hash import crash  # type: ignore
 
 from buckaroo.file_cache.mp_timeout_decorator import (
-    mp_simple, mp_sleep1, mp_crash_exit, mp_polars_longread, mp_polars_crash,
     TimeoutException, ExecutionFailed, mp_timeout
+)
+
+from .mp_module_funcs import ( mp_simple, mp_sleep1, mp_crash_exit, mp_polars_longread, mp_polars_crash
 )
